Combining Custom Converters and Review comments rework
